Separated classes into packages by animal type (walking, swimming, slithering)
