+++ swimming/request.py
@@ -1,0 +1,35 @@
+from datetime import date
+
+class Fish():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
+
+class Seal():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
+class Shark():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
+
+class Eel():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
+
+class Turtle():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
