+++ walking/__init__.py
@@ -1,0 +1,1 @@
+from .request import (Llama, Donkey, Goat, Horse, Chicken)
