+++ walking/request.py
@@ -1,0 +1,40 @@
+from datetime import date
+
+class Llama():
+
+    def __init__(self, name, species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
+class Donkey():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
+class Goat():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
+class Horse():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
+class Chicken():
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
